Deprecate custom `JsonPatch` implementation
